10527 2112 input bool로 바꿔야돼

+++ ManagerApplication/common_header.h
@@ -11,6 +11,6 @@ enum FUNCTION_RETURN
 	FUNCTION_NULL = -3,
 };
 
-typedef int8_t FNRETURN;
+typedef int16_t FNRETURN;
 
 FNRETURN inputInteger();
